refactor: replace EventTranslatorConfigFactory with Jackson XmlMapper — Phase 4
- Move EventTranslatorConfig interface from opennms-config to opennms-config-api
  (same package — no import changes needed)
- Add model constructor to EventTranslatorConfigFactory accepting pre-deserialized
  EventTranslatorConfiguration + DataSource, bypassing JAXB and file I/O
- Update EventTranslatorBootConfiguration to load translator-configuration.xml
  via Jackson XmlMapper + JaxbAnnotationModule (consistent with Trapd, Syslogd,
  Discovery patterns)
- Widen bean return type from concrete factory to EventTranslatorConfig interface
- features/event-translator now depends on opennms-config-api (interface only),
  decoupled from opennms-config
- Delete dead Karaf-era JMX wrapper (jmx/EventTranslator, EventTranslatorMBean)
